% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_roost_tbl.R
\docType{data}
\name{bat_roost_tbl}
\alias{bat_roost_tbl}
\title{Max counts of bat roosts of different species across the UK}
\format{
A tibble with 3,054 rows and 6 variables:
\describe{
  \item{Species}{chr Bat species}
  \item{RoostCode}{fct Unique Roost code} 
  \item{MaxCountInAnyYear}{dbl Max yearly bat population count of roost}
  \item{County}{chr Country roost is located in}
  \item{Regio}{chr Region roost is located in}
  \item{Country}{chr Country roost is located in}
}
}
\source{
\url{https://www.bats.org.uk/our-work/national-bat-monitoring-programme/reports/nbmp-annual-report}
}
\usage{
bat_roost_tbl
}
\description{
The dataset was produced by The National Bat Monitoring Programme (NBMP) in the UK
Contains National Bat Monitoring Programme (NBMP) data © copyright and database right Bat Conservation Trust and the Joint Nature Conservation Committee, 2025.
}
\keyword{datasets}
