% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_roost_wide_tbl.R
\docType{data}
\name{bat_roost_wide_tbl}
\alias{bat_roost_wide_tbl}
\title{Mean max counts of bat roosts of different species across the UK}
\format{
A tibble with 128 rows and 20 variables:
\describe{
  \item{Area}{chr County;Region;Country}
  \item{The rest}{dbl The mean max count in any year of the species across all the investigated roosts across the area} 
}
}
\source{
\url{https://www.bats.org.uk/our-work/national-bat-monitoring-programme/reports/nbmp-annual-report}
}
\usage{
bat_roost_wide_tbl
}
\description{
The dataset was produced by The National Bat Monitoring Programme (NBMP) in the UK
Contains National Bat Monitoring Programme (NBMP) data © copyright and database right Bat Conservation Trust and the Joint Nature Conservation Committee, 2025.
}
\keyword{datasets}
