% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries_long_tbl.R
\docType{data}
\name{fisheries_long_tbl}
\alias{fisheries_long_tbl}
\title{Data of fish and overfishing}
\format{
A tibble with 14,674 rows and 3 variables:
\describe{
  \item{Entitity}{chr Country or region}
  \item{Year}{dbl Gregorian calendar year in AD} 
  \item{Metric_tons}{dbl Captured production in metric tons}
}
}
\source{
\url{https://www.kaggle.com/datasets/sergegeukjian/fish-and-overfishing?select=capture-fisheries-vs-aquaculture.csv}
}
\usage{
fisheries_long_tbl
}
\description{
A long formatted tibble containing information on capture fishery production on various entities from 1960 to 2018
}
\keyword{datasets}
