% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries_wide_tbl.R
\docType{data}
\name{fisheries_wide_tbl}
\alias{fisheries_wide_tbl}
\title{Data of fish and overfishing}
\format{
A tibble with 264 rows and 60 variables:
\describe{
  \item{Entity}{chr Country or region}
  \item{1960-2018}{dbl The year's captured production in metric tons} 
}
}
\source{
\url{https://www.kaggle.com/datasets/sergegeukjian/fish-and-overfishing?select=capture-fisheries-vs-aquaculture.csv}
}
\usage{
fisheries_wide_tbl
}
\description{
A wide formatted tibble containing information on capture fishery production on various entities from 1960 to 2018
}
\keyword{datasets}
