% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/living_planet_wide_tbl.R
\docType{data}
\name{living_planet_wide_tbl}
\alias{living_planet_wide_tbl}
\title{Average change in population size of of studied animal population}
\format{
A tibble with 342 rows and 5 variables:
\describe{
  \item{Region}{fct Region or group}
  \item{1970-2018}{dbl The year's average living planet index}
}
}
\source{
\url{https://www.kaggle.com/datasets/elmoallistair/living-planet-index}
}
\usage{
living_planet_wide_tbl
}
\description{
The Living Planet Index (LPI) is a measure of the state of global biological diversity based on population trends of vertebrate species from around the world. The index value is measured relative to species' populations in 1970 (i.e. 1970 = 1).
}
\keyword{datasets}
