% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrtibbles-package.R
\docType{package}
\name{mgrtibbles-package}
\alias{mgrtibbles}
\alias{mgrtibbles-package}
\title{mgrtibbles: R tibbles containing various datasets}
\description{
Tibbles for examples in teaching/learning.
}
\author{
\strong{Maintainer}: Matthew Gemmell \email{matthewrgemmell@gmail.com}

}
\keyword{internal}
