% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_pop_wide_tbl.R
\docType{data}
\name{world_pop_wide_tbl}
\alias{world_pop_wide_tbl}
\title{Global Headcount: World Population Dataset by Country/Territory}
\format{
A tibble with 234 rows and 13 variables:
\describe{
  \item{CCA3}{fct 3 Digit Country/Territories Code}
  \item{Country/Territories}{fct Name of the Country/Territories} 
  \item{Capital}{chr Name of the Capital}
  \item{Continent}{fct Name of the Continent}
  \item{2022-1970}{dbl Population of the Country/Territory in the denoted year}
  \item{Area_km2}{dbl Area size of the Country/Territories in square kilometer}
}
}
\source{
\url{https://www.kaggle.com/datasets/iamsouravbanerjee/world-population-dataset}
}
\usage{
world_pop_wide_tbl
}
\description{
Global Headcount: World Population Dataset by Country/Territory
}
\keyword{datasets}
